+++ CHANGELOG.md
@@ -3,6 +3,19 @@
 All notable changes to the Zowe Launcher package will be documented in this file.
 This repo is part of the app-server Zowe Component, and the change logs here may appear on Zowe.org in that section.
 
+## 3.6.0
+- Bugfix: `zowe.sysMessages` are used after the validation ([#208](https://github.com/zowe/launcher/pull/208))
+- Enhancement: regex replaced by internal routine ([#206](https://github.com/zowe/launcher/pull/206))
+- Bugfix: Component names must fit the component list ([#195](https://github.com/zowe/launcher/pull/195))
+- Enhancement: More restrictive permissions ([#212](https://github.com/zowe/launcher/pull/212))
+
+## 3.5.0
+- Bugfix: Escape backslash, backtick and dollar sign when used in environment variable ([#168](https://github.com/zowe/launcher/pull/168))
+- Bugfix: Check the array size of `zowe.launcher.restartIntervals` ([#187](https://github.com/zowe/launcher/pull/187))
+- Bugfix: Check the HA instance name if followed by any additional parameters ([#181](https://github.com/zowe/launcher/pull/181))
+- Bugfix: Use `wtoMessage` directly when the message is triggered from `zowe.sysMessages` ([#189](https://github.com/zowe/launcher/pull/189))
+- Enhancement: All error messages are printed to `SYSPRINT` ([#177](https://github.com/zowe/launcher/pull/177))
+
 ## 3.4.0
 - Enhancement: Message `ZWEL0021I` includes the version of launcher ([#167](https://github.com/zowe/launcher/pull/167))